[nfc] [cxx-interop] Move function template specialization logic into clang-module-loader.
