+++ lib/Sema/CSApply.cpp
@@ -109,368 +109,6 @@ Solution::computeSubstitutions(GenericSignature sig,
                               lookupConformanceFn);
 }
 
-// On Windows and 32-bit platforms we need to force "Int" to actually be
-// re-imported as "Int." This is needed because otherwise, we cannot round-trip
-// "Int" and "UInt". For example, on Windows, "Int" will be imported into C++ as
-// "long long" and then back into Swift as "Int64" not "Int."
-static ValueDecl *rewriteIntegerTypes(SubstitutionMap subst, ValueDecl *oldDecl,
-                                      AbstractFunctionDecl *newDecl) {
-  auto originalFnSubst = cast<AbstractFunctionDecl>(oldDecl)
-                             ->getInterfaceType()
-                             ->getAs<GenericFunctionType>()
-                             ->substGenericArgs(subst);
-  // The constructor type is a function type as follows:
-  //   (CType.Type) -> (Generic) -> CType
-  // And a method's function type is as follows:
-  //   (inout CType) -> (Generic) -> Void
-  // In either case, we only want the result of that function type because that
-  // is the function type with the generic params that need to be substituted:
-  //   (Generic) -> CType
-  if (isa<ConstructorDecl>(oldDecl) || oldDecl->isInstanceMember() ||
-      oldDecl->isStatic())
-    originalFnSubst = cast<FunctionType>(originalFnSubst->getResult().getPointer());
-
-  SmallVector<ParamDecl *, 4> fixedParameters;
-  unsigned parameterIndex = 0;
-  for (auto *newFnParam : *newDecl->getParameters()) {
-    // If the user substituted this param with an (U)Int, use (U)Int.
-    auto substParamType =
-        originalFnSubst->getParams()[parameterIndex].getParameterType();
-    if (substParamType->isEqual(newDecl->getASTContext().getIntType()) ||
-        substParamType->isEqual(newDecl->getASTContext().getUIntType())) {
-      auto intParam =
-          ParamDecl::cloneWithoutType(newDecl->getASTContext(), newFnParam);
-      intParam->setInterfaceType(substParamType);
-      fixedParameters.push_back(intParam);
-    } else {
-      fixedParameters.push_back(newFnParam);
-    }
-    parameterIndex++;
-  }
-
-  auto fixedParams =
-      ParameterList::create(newDecl->getASTContext(), fixedParameters);
-  newDecl->setParameters(fixedParams);
-
-  // Now fix the result type:
-  if (originalFnSubst->getResult()->isEqual(
-          newDecl->getASTContext().getIntType()) ||
-      originalFnSubst->getResult()->isEqual(
-          newDecl->getASTContext().getUIntType())) {
-    // Constructors don't have a result.
-    if (auto func = dyn_cast<FuncDecl>(newDecl)) {
-      // We have to rebuild the whole function.
-      auto newFnDecl = FuncDecl::createImported(
-          func->getASTContext(), func->getNameLoc(),
-          func->getName(), func->getNameLoc(),
-          func->hasAsync(), func->hasThrows(),
-          fixedParams, originalFnSubst->getResult(),
-          /*genericParams=*/nullptr, func->getDeclContext(), newDecl->getClangDecl());
-      if (func->isStatic()) newFnDecl->setStatic();
-      if (func->isImportAsStaticMember()) newFnDecl->setImportAsStaticMember();
-      if (func->getImportAsMemberStatus().isInstance()) {
-        newFnDecl->setSelfAccessKind(func->getSelfAccessKind());
-        newFnDecl->setSelfIndex(func->getSelfIndex());
-      }
-
-      return newFnDecl;
-    }
-  }
-
-  return newDecl;
-}
-
-static Expr *createSelfExpr(FuncDecl *fnDecl) {
-  ASTContext &ctx = fnDecl->getASTContext();
-
-  auto selfDecl = fnDecl->getImplicitSelfDecl();
-  auto selfRefExpr = new (ctx) DeclRefExpr(selfDecl, DeclNameLoc(),
-                                           /*implicit*/ true);
-
-  if (!fnDecl->isMutating()) {
-    selfRefExpr->setType(selfDecl->getInterfaceType());
-    return selfRefExpr;
-  }
-  selfRefExpr->setType(LValueType::get(selfDecl->getInterfaceType()));
-
-  auto inoutSelfExpr = new (ctx) InOutExpr(
-      SourceLoc(), selfRefExpr,
-      fnDecl->mapTypeIntoContext(selfDecl->getValueInterfaceType()),
-      /*isImplicit*/ true);
-  inoutSelfExpr->setType(InOutType::get(selfDecl->getInterfaceType()));
-  return inoutSelfExpr;
-}
-
-// Synthesize a thunk body for the function created in
-// "addThunkForDependentTypes". This will just cast all params and forward them
-// along to the specialized function. It will also cast the result before
-// returning it.
-static std::pair<BraceStmt *, bool>
-synthesizeDependentTypeThunkParamForwarding(AbstractFunctionDecl *afd, void *context) {
-  ASTContext &ctx = afd->getASTContext();
-
-  auto thunkDecl = cast<FuncDecl>(afd);
-  auto specializedFuncDecl = static_cast<FuncDecl *>(context);
-
-  SmallVector<Argument, 8> forwardingParams;
-  unsigned paramIndex = 0;
-  for (auto param : *thunkDecl->getParameters()) {
-    if (isa<MetatypeType>(param->getType().getPointer())) {
-      paramIndex++;
-      continue;
-    }
-
-    Expr *paramRefExpr = new (ctx) DeclRefExpr(param, DeclNameLoc(),
-                                               /*Implicit=*/true);
-    paramRefExpr->setType(param->getType());
-
-    if (param->isInOut()) {
-      paramRefExpr->setType(LValueType::get(param->getType()));
-
-      paramRefExpr = new (ctx) InOutExpr(
-          SourceLoc(), paramRefExpr, param->getType(), /*isImplicit*/ true);
-      paramRefExpr->setType(InOutType::get(param->getType()));
-    }
-
-    auto specParamTy = specializedFuncDecl->getParameters()->get(paramIndex)->getType();
-
-    Expr *argExpr = nullptr;
-    if (specParamTy->isEqual(param->getType())) {
-      argExpr = paramRefExpr;
-    } else {
-      argExpr = ForcedCheckedCastExpr::createImplicit(ctx, paramRefExpr,
-                                                      specParamTy);
-    }
-
-    forwardingParams.push_back(Argument(SourceLoc(), Identifier(), argExpr));
-    paramIndex++;
-  }
-
-  Expr *specializedFuncDeclRef = new (ctx) DeclRefExpr(ConcreteDeclRef(specializedFuncDecl),
-                                                       DeclNameLoc(), true);
-  specializedFuncDeclRef->setType(specializedFuncDecl->getInterfaceType());
-
-  if (specializedFuncDecl->isInstanceMember()) {
-    auto selfExpr = createSelfExpr(thunkDecl);
-    auto *memberCall = DotSyntaxCallExpr::create(ctx, specializedFuncDeclRef, SourceLoc(), selfExpr);
-    memberCall->setThrows(false);
-    auto resultType = specializedFuncDecl->getInterfaceType()->getAs<FunctionType>()->getResult();
-    specializedFuncDeclRef = memberCall;
-    specializedFuncDeclRef->setType(resultType);
-  } else if (specializedFuncDecl->isStatic()) {
-    auto resultType = specializedFuncDecl->getInterfaceType()->getAs<FunctionType>()->getResult();
-    auto selfType = cast<NominalTypeDecl>(thunkDecl->getDeclContext()->getAsDecl())->getDeclaredInterfaceType();
-    auto selfTypeExpr = TypeExpr::createImplicit(selfType, ctx);
-    auto *memberCall = DotSyntaxCallExpr::create(ctx, specializedFuncDeclRef, SourceLoc(), selfTypeExpr);
-    memberCall->setThrows(false);
-    specializedFuncDeclRef = memberCall;
-    specializedFuncDeclRef->setType(resultType);
-  }
-
-  auto argList = ArgumentList::createImplicit(ctx, forwardingParams);
-  auto *specializedFuncCallExpr = CallExpr::createImplicit(ctx, specializedFuncDeclRef, argList);
-  specializedFuncCallExpr->setType(specializedFuncDecl->getResultInterfaceType());
-  specializedFuncCallExpr->setThrows(false);
-
-  Expr *resultExpr = nullptr;
-  if (specializedFuncCallExpr->getType()->isEqual(thunkDecl->getResultInterfaceType())) {
-    resultExpr = specializedFuncCallExpr;
-  } else {
-    resultExpr = ForcedCheckedCastExpr::createImplicit(
-        ctx, specializedFuncCallExpr, thunkDecl->getResultInterfaceType());
-  }
-
-  auto returnStmt = new (ctx)
-      ReturnStmt(SourceLoc(), resultExpr, /*implicit=*/true);
-  auto body = BraceStmt::create(ctx, SourceLoc(), {returnStmt}, SourceLoc(),
-                                /*implicit=*/true);
-  return {body, /*isTypeChecked=*/true};
-}
-
-// Create a thunk to map functions with dependent types to their specialized
-// version. For example, create a thunk with type (Any) -> Any to wrap a
-// specialized function template with type (Dependent<T>) -> Dependent<T>.
-static ValueDecl *addThunkForDependentTypes(FuncDecl *oldDecl,
-                                            FuncDecl *newDecl) {
-  bool updatedAnyParams = false;
-
-  SmallVector<ParamDecl *, 4> fixedParameters;
-  unsigned parameterIndex = 0;
-  for (auto *newFnParam : *newDecl->getParameters()) {
-    // If the un-specialized function had a parameter with type "Any" preserve
-    // that parameter. Otherwise, use the new function parameter.
-    auto oldParamType = oldDecl->getParameters()->get(parameterIndex)->getType();
-    if (oldParamType->isEqual(newDecl->getASTContext().TheAnyType)) {
-      updatedAnyParams = true;
-      auto newParam =
-          ParamDecl::cloneWithoutType(newDecl->getASTContext(), newFnParam);
-      newParam->setInterfaceType(oldParamType);
-      fixedParameters.push_back(newParam);
-    } else {
-      fixedParameters.push_back(newFnParam);
-    }
-    parameterIndex++;
-  }
-
-  // If we don't need this thunk, bail out.
-  if (!updatedAnyParams &&
-      !oldDecl->getResultInterfaceType()->isEqual(
-          oldDecl->getASTContext().TheAnyType))
-    return newDecl;
-
-  auto fixedParams =
-      ParameterList::create(newDecl->getASTContext(), fixedParameters);
-
-  Type fixedResultType;
-  if (oldDecl->getResultInterfaceType()->isEqual(
-          oldDecl->getASTContext().TheAnyType))
-      fixedResultType = oldDecl->getASTContext().TheAnyType;
-  else
-    fixedResultType = newDecl->getResultInterfaceType();
-
-  // We have to rebuild the whole function.
-  auto newFnDecl = FuncDecl::createImplicit(
-      newDecl->getASTContext(), newDecl->getStaticSpelling(),
-      newDecl->getName(), newDecl->getNameLoc(), newDecl->hasAsync(),
-      newDecl->hasThrows(), /*genericParams=*/nullptr, fixedParams,
-      fixedResultType, newDecl->getDeclContext());
-  newFnDecl->copyFormalAccessFrom(newDecl);
-  newFnDecl->setBodySynthesizer(synthesizeDependentTypeThunkParamForwarding, newDecl);
-  newFnDecl->setSelfAccessKind(newDecl->getSelfAccessKind());
-  if (newDecl->isStatic()) newFnDecl->setStatic();
-  newFnDecl->getAttrs().add(
-      new (newDecl->getASTContext()) TransparentAttr(/*IsImplicit=*/true));
-  return newFnDecl;
-}
-
-// Synthesizes the body of a thunk that takes extra metatype arguments and
-// skips over them to forward them along to the FuncDecl contained by context.
-// This is used when importing a C++ templated function where the template params
-// are not used in the function signature. We supply the type params as explicit
-// metatype arguments to aid in typechecking, but they shouldn't be forwarded to
-// the corresponding C++ function.
-static std::pair<BraceStmt *, bool>
-synthesizeForwardingThunkBody(AbstractFunctionDecl *afd, void *context) {
-  ASTContext &ctx = afd->getASTContext();
-
-  auto thunkDecl = cast<FuncDecl>(afd);
-  auto specializedFuncDecl = static_cast<FuncDecl *>(context);
-
-  SmallVector<Argument, 8> forwardingParams;
-  for (auto param : *thunkDecl->getParameters()) {
-    if (isa<MetatypeType>(param->getType().getPointer())) {
-      continue;
-    }
-    Expr *paramRefExpr = new (ctx) DeclRefExpr(param, DeclNameLoc(),
-                                               /*Implicit=*/true);
-    paramRefExpr->setType(param->getType());
-
-    if (param->isInOut()) {
-      paramRefExpr->setType(LValueType::get(param->getType()));
-
-      paramRefExpr = new (ctx) InOutExpr(
-          SourceLoc(), paramRefExpr, param->getType(), /*isImplicit*/ true);
-      paramRefExpr->setType(InOutType::get(param->getType()));
-    }
-
-    forwardingParams.push_back(Argument(SourceLoc(), Identifier(), paramRefExpr));
-  }
-
-  Expr *specializedFuncDeclRef = new (ctx) DeclRefExpr(ConcreteDeclRef(specializedFuncDecl),
-                                                DeclNameLoc(), true);
-  specializedFuncDeclRef->setType(specializedFuncDecl->getInterfaceType());
-
-  if (specializedFuncDecl->isInstanceMember()) {
-    auto selfExpr = createSelfExpr(thunkDecl);
-    auto *memberCall = DotSyntaxCallExpr::create(ctx, specializedFuncDeclRef, SourceLoc(), selfExpr);
-    memberCall->setThrows(false);
-    auto resultType = specializedFuncDecl->getInterfaceType()->getAs<FunctionType>()->getResult();
-    specializedFuncDeclRef = memberCall;
-    specializedFuncDeclRef->setType(resultType);
-  } else if (specializedFuncDecl->isStatic()) {
-    auto resultType = specializedFuncDecl->getInterfaceType()->getAs<FunctionType>()->getResult();
-    auto selfType = cast<NominalTypeDecl>(thunkDecl->getDeclContext()->getAsDecl())->getDeclaredInterfaceType();
-    auto selfTypeExpr = TypeExpr::createImplicit(selfType, ctx);
-    auto *memberCall = DotSyntaxCallExpr::create(ctx, specializedFuncDeclRef, SourceLoc(), selfTypeExpr);
-    memberCall->setThrows(false);
-    specializedFuncDeclRef = memberCall;
-    specializedFuncDeclRef->setType(resultType);
-  }
-
-  auto argList = ArgumentList::createImplicit(ctx, forwardingParams);
-  auto *specializedFuncCallExpr = CallExpr::createImplicit(ctx, specializedFuncDeclRef, argList);
-  specializedFuncCallExpr->setType(thunkDecl->getResultInterfaceType());
-  specializedFuncCallExpr->setThrows(false);
-
-  auto returnStmt = new (ctx) ReturnStmt(SourceLoc(), specializedFuncCallExpr,
-                                        /*implicit=*/true);
-
-  auto body = BraceStmt::create(ctx, SourceLoc(), {returnStmt}, SourceLoc(),
-                               /*implicit=*/true);
-  return {body, /*isTypeChecked=*/true};
-}
-
-static ValueDecl *generateThunkForExtraMetatypes(SubstitutionMap subst,
-                                                 FuncDecl *oldDecl,
-                                                 FuncDecl *newDecl) {
-  // We added additional metatype parameters to aid template
-  // specialization, which are no longer now that we've specialized
-  // this function. Create a thunk that only forwards the original
-  // parameters along to the clang function.
-  SmallVector<ParamDecl *, 4> newParams;
-
-  for (auto param : *newDecl->getParameters()) {
-    auto *newParamDecl = ParamDecl::clone(newDecl->getASTContext(), param);
-    newParams.push_back(newParamDecl);
-  }
-
-  auto originalFnSubst = cast<AbstractFunctionDecl>(oldDecl)
-                             ->getInterfaceType()
-                             ->getAs<GenericFunctionType>()
-                             ->substGenericArgs(subst);
-  // The constructor type is a function type as follows:
-  //   (CType.Type) -> (Generic) -> CType
-  // And a method's function type is as follows:
-  //   (inout CType) -> (Generic) -> Void
-  // In either case, we only want the result of that function type because that
-  // is the function type with the generic params that need to be substituted:
-  //   (Generic) -> CType
-  if (isa<ConstructorDecl>(oldDecl) || oldDecl->isInstanceMember() ||
-      oldDecl->isStatic())
-    originalFnSubst = cast<FunctionType>(originalFnSubst->getResult().getPointer());
-
-  for (auto paramTy : originalFnSubst->getParams()) {
-    if (!paramTy.getPlainType()->is<MetatypeType>())
-      continue;
-
-    auto dc = newDecl->getDeclContext();
-    auto paramVarDecl =
-        new (newDecl->getASTContext()) ParamDecl(
-            SourceLoc(), SourceLoc(), Identifier(), SourceLoc(),
-            newDecl->getASTContext().getIdentifier("_"), dc);
-    paramVarDecl->setInterfaceType(paramTy.getPlainType());
-    paramVarDecl->setSpecifier(ParamSpecifier::Default);
-    newParams.push_back(paramVarDecl);
-  }
-
-  auto *newParamList =
-      ParameterList::create(newDecl->getASTContext(), SourceLoc(), newParams, SourceLoc());
-
-  auto thunk = FuncDecl::createImplicit(
-      newDecl->getASTContext(), newDecl->getStaticSpelling(), oldDecl->getName(),
-      newDecl->getNameLoc(), newDecl->hasAsync(), newDecl->hasThrows(),
-      /*genericParams=*/nullptr, newParamList,
-      newDecl->getResultInterfaceType(), newDecl->getDeclContext());
-  thunk->copyFormalAccessFrom(newDecl);
-  thunk->setBodySynthesizer(synthesizeForwardingThunkBody, newDecl);
-  thunk->setSelfAccessKind(newDecl->getSelfAccessKind());
-  if (newDecl->isStatic()) thunk->setStatic();
-  thunk->getAttrs().add(
-      new (newDecl->getASTContext()) TransparentAttr(/*IsImplicit=*/true));
-
-  return thunk;
-}
-
 // Lazily instantiate function definitions for class template specializations.
 // Members of a class template specialization will be instantiated here (not
 // when imported). If this method has already be instantiated, then this is a
@@ -488,48 +126,6 @@ static void maybeInstantiateCXXMethodDefinition(ValueDecl *decl) {
   }
 }
 
-static ConcreteDeclRef getCXXFunctionTemplateSpecialization(SubstitutionMap subst,
-                                                            ValueDecl *decl) {
-  assert(isa<clang::FunctionTemplateDecl>(decl->getClangDecl()) &&
-      "This API should only be used with function templates.");
-
-  auto *newFn =
-      decl->getASTContext()
-          .getClangModuleLoader()
-          ->instantiateCXXFunctionTemplate(
-              decl->getASTContext(),
-              const_cast<clang::FunctionTemplateDecl *>(
-                  cast<clang::FunctionTemplateDecl>(decl->getClangDecl())),
-              subst);
-  // We failed to specialize this function template. The compiler is going to
-  // exit soon. Return something valid in the meantime.
-  if (!newFn)
-    return ConcreteDeclRef(decl);
-
-  auto newDecl = cast_or_null<ValueDecl>(
-      decl->getASTContext().getClangModuleLoader()->importDeclDirectly(
-          newFn));
-
-  if (auto fn = dyn_cast<AbstractFunctionDecl>(newDecl)) {
-    if (!subst.empty()) {
-      newDecl = rewriteIntegerTypes(subst, decl, fn);
-    }
-  }
-
-  if (auto fn = dyn_cast<FuncDecl>(decl)) {
-    newDecl = addThunkForDependentTypes(fn, cast<FuncDecl>(newDecl));
-  }
-
-  if (auto fn = dyn_cast<FuncDecl>(decl)) {
-    if (newFn->getNumParams() != fn->getParameters()->size()) {
-      newDecl = generateThunkForExtraMetatypes(subst, fn,
-                                               cast<FuncDecl>(newDecl));
-    }
-  }
-
-  return ConcreteDeclRef(newDecl);
-}
-
 ConcreteDeclRef
 Solution::resolveConcreteDeclRef(ValueDecl *decl,
                                  ConstraintLocator *locator) const {
@@ -545,7 +141,8 @@ Solution::resolveConcreteDeclRef(ValueDecl *decl,
   // If this is a C++ function template, get it's specialization for the given
   // substitution map and update the decl accordingly.
   if (isa_and_nonnull<clang::FunctionTemplateDecl>(decl->getClangDecl())) {
-    return getCXXFunctionTemplateSpecialization(subst, decl);
+    auto moduleLoader = decl->getASTContext().getClangModuleLoader();
+    return moduleLoader->getCXXFunctionTemplateSpecialization(subst, decl);
   }
 
   return ConcreteDeclRef(decl, subst);
